backend/requirements.txt
langchain
langchain-openai
fastapi
uvicorn
python-dotenv
